updated saving client sample
