+++ SavingDashboards-Client/client/index.html
@@ -22,55 +22,53 @@
     <script src="https://cdn.jsdelivr.net/npm/jquery@3.6.0/dist/jquery.min.js"></script>
     <script src="https://cdn.quilljs.com/1.3.6/quill.min.js"></script>
     <script src="https://unpkg.com/dayjs@1.8.21/dayjs.min.js"></script>
-    <script src="https://dl.revealbi.io/reveal/libs/1.1.7/infragistics.reveal.js"></script>
+    <script src="https://dl.revealbi.io/reveal/libs/1.3.0/infragistics.reveal.js"></script>
 
     <script type="text/javascript">
         //set this to your server url
-        $.ig.RevealSdkSettings.setBaseUrl("https://localhost:7111/");
+        $.ig.RevealSdkSettings.setBaseUrl("http://localhost:5111/");
 
-        $.ig.RevealSdkSettings.ensureFontsLoadedAsync().then(() => {
-            $.ig.RVDashboard.loadDashboard("Sales", (dashboard) => {
-                var revealView = new $.ig.RevealView("#revealView");
-                revealView.serverSideSave = false;
-                revealView.dashboard = dashboard;
+        $.ig.RVDashboard.loadDashboard("Sales").then(dashboard => {
+            var revealView = new $.ig.RevealView("#revealView");
+            revealView.serverSideSave = false;
+            revealView.dashboard = dashboard;
 
-                revealView.onSave = (rv, args) => {
-                    if (args.saveAs) {
-                        var newName = prompt("Please enter the dashboard name");
-                        isDuplicateName(newName).then(isDuplicate => {
-                            if (isDuplicate === "true") {
-                                if (!window.confirm("A dashboard with name: " + newName + " already exists. Do you want to override it?")) {
-                                    args.saveFinished();
-                                    return;
-                                }
+            revealView.onSave = (rv, args) => {
+                if (args.saveAs) {
+                    var newName = prompt("Please enter the dashboard name");
+                    isDuplicateName(newName).then(isDuplicate => {
+                        if (isDuplicate === "true") {
+                            if (!window.confirm("A dashboard with name: " + newName + " already exists. Do you want to override it?")) {
+                                args.saveFinished();
+                                return;
                             }
+                        }
 
-                            args.dashboardId = args.name = newName;
-                            args.serializeWithNewName(newName, bytes => {
-                                this.saveDashboard(newName, bytes, true).then(() => {
-                                    args.saveFinished();
-                                });
-                            });
-                        });
-                    }
-                    else {
-                        args.serialize(bytes => {
-                            this.saveDashboard(args.name, bytes).then(() => {
+                        args.dashboardId = args.name = newName;
+                        args.serializeWithNewName(newName, bytes => {
+                            this.saveDashboard(newName, bytes, true).then(() => {
                                 args.saveFinished();
                             });
                         });
-                    }
-                };
-            });
+                    });
+                }
+                else {
+                    args.serialize(bytes => {
+                        this.saveDashboard(args.name, bytes).then(() => {
+                            args.saveFinished();
+                        });
+                    });
+                }
+            };
         });
 
         function isDuplicateName(name) {
-            return fetch(`https://localhost:7111/isduplicatename/${name}`).then(resp => resp.text());
+            return fetch(`http://localhost:5111/isduplicatename/${name}`).then(resp => resp.text());
         }
 
         function saveDashboard(name, bytes, isSaveAs = false) {
 
-            let url = `https://localhost:7111/dashboards/${name}`;
+            let url = `http://localhost:5111/dashboards/${name}`;
             let params = {
                 body: bytes,
                 method: "PUT"
